+++ test/cli.js
@@ -1,95 +1,116 @@
 var bag = require('bagofholding'),
-  sandbox = require('sandboxed-module'),
-  should = require('should'),
-  checks, mocks,
-  cli;
+  buster = require('buster'),
+  cli = require('../lib/cli'),
+  Breaker = require('../lib/breaker');
 
-describe('cli', function () {
-
-  function create(checks, mocks) {
-    return sandbox.require('../lib/cli', {
-      requires: {
-        bagofholding: {
-          cli: {
-            exit: bag.cli.exit,
-            parse: function (commands, dir) {
-              checks.bag_parse_commands = commands;
-              checks.bag_parse_dir = dir;
-            }
-          }
-        },
-        './breaker': function () {
-          return {
-            init: function (exit) {
-              checks.breaker_init_exit = exit;
-            },
-            format: function (type, exit) {
-              checks.breaker_format_type = type;
-              checks.breaker_format_exit = exit;
-            },
-            ssh: function (command, labels, exit) {
-              checks.breaker_ssh_command = command;
-              checks.breaker_ssh_labels = labels;
-              checks.breaker_ssh_exit = exit;
-            }
-          };
-        }
-      },
-      globals: {
-        process: bag.mock.process(checks, mocks)
-      }
-    });
-  }
-
-  beforeEach(function () {
-    checks = {};
-    mocks = {
-      process_cwd: '/somedir/breaker'
+buster.testCase('cli - exec', {
+  'should contain commands with actions': function (done) {
+    var mockCommand = function (base, actions) {
+      assert.defined(base);
+      assert.defined(actions.commands.init.action);
+      assert.defined(actions.commands.format.action);
+      assert.defined(actions.commands.ssh.action);
+      done();
     };
-    cli = create(checks, mocks);
+    this.stub(bag, 'cli', { command: mockCommand });
     cli.exec();
-  });
-
-  describe('exec', function () {
-
-    it('should contain init command and delegate to breaker init when exec is called', function () {
-      checks.bag_parse_commands.init.desc.should.equal('Create sample Breaker hosts file');
-      checks.bag_parse_commands.init.action();
-      checks.breaker_init_exit.should.be.a('function');
-    });
-
-    it('should contain format command with default type and delegate to breaker format when exec is called', function () {
-      checks.bag_parse_commands.format.desc.should.equal('Format hosts info into specific type');
-      checks.bag_parse_commands.format.options[0].arg.should.equal('-t, --type <type>');
-      checks.bag_parse_commands.format.options[0].desc.should.equal('Format type');
-      checks.bag_parse_commands.format.action({});
-      checks.breaker_format_type.should.equal('sshconfig');
-      checks.breaker_format_exit.should.be.a('function');
-    });
-
-    it('should contain format command with type when args contain specified type', function () {
-      checks.bag_parse_commands.format.desc.should.equal('Format hosts info into specific type');
-      checks.bag_parse_commands.format.action({ type: 'clusterssh' });
-      checks.breaker_format_type.should.equal('clusterssh');
-      checks.breaker_format_exit.should.be.a('function');
-    });
-
-    it('should contain ssh command and delegate to breaker ssh when exec is called', function () {
-      checks.bag_parse_commands.ssh.desc.should.equal('Remotely execute shell command on selected configured hosts in parallel via SSH');
-      checks.bag_parse_commands.ssh.options[0].arg.should.equal('-l, --labels <labels>');
-      checks.bag_parse_commands.ssh.options[0].desc.should.equal('Comma separated labels');
-      checks.bag_parse_commands.ssh.action('df -kh;', { labels: 'foo,bar' });
-      checks.breaker_ssh_command.should.equal('df -kh;');
-      checks.breaker_ssh_labels.length.should.equal(2);
-      checks.breaker_ssh_labels[0].should.equal('foo');
-      checks.breaker_ssh_labels[1].should.equal('bar');
-      checks.breaker_ssh_exit.should.be.a('function');
-    });
-
-    it('should pass empty array labels when labels flag is not specified', function () {
-      checks.bag_parse_commands.ssh.action('df -kh;', {});
-      checks.breaker_ssh_labels.length.should.equal(0);
-    });
-  });
+  }
 });
- 
+
+buster.testCase('cli - init', {
+  setUp: function () {
+    this.mockConsole = this.mock(console);
+  },
+  'should contain init command and delegate to breaker init when exec is called': function (done) {
+    this.mockConsole.expects('log').once().withExactArgs('Creating sample Breaker hosts file: .breaker.json');
+    this.stub(bag, 'cli', {
+      command: function (base, actions) {
+        actions.commands.init.action();
+      },
+      exit: bag.cli.exit
+    });
+    this.stub(Breaker.prototype, 'init', function (cb) {
+      assert.equals(typeof cb, 'function');
+      done();
+    });
+    cli.exec();
+  }
+});
+
+buster.testCase('cli - format', {
+  'should contain format command and delegate to breaker format when exec is called': function (done) {
+    this.stub(bag, 'cli', {
+      command: function (base, actions) {
+        actions.commands.format.action();
+      },
+      exit: bag.cli.exit
+    });
+    this.stub(Breaker.prototype, 'format', function (type, cb) {
+      assert.equals(type, 'sshconfig');
+      assert.equals(typeof cb, 'function');
+      done();
+    });
+    cli.exec();
+  },
+  'should use custom type when passed via args': function (done) {
+    this.stub(bag, 'cli', {
+      command: function (base, actions) {
+        actions.commands.format.action({ type: 'clusterssh' });
+      },
+      exit: bag.cli.exit
+    });
+    this.stub(Breaker.prototype, 'format', function (type, cb) {
+      assert.equals(type, 'clusterssh');
+      assert.equals(typeof cb, 'function');
+      done();
+    });
+    cli.exec();
+  },
+  'should pass labels array via constructor when argument labels are provided': function (done) {
+    this.stub(bag, 'cli', {
+      command: function (base, actions) {
+        actions.commands.format.action({ labels: 'foo,bar' });
+      },
+      exit: bag.cli.exit
+    });
+    this.stub(Breaker.prototype, 'format', function (type, cb) {
+      assert.equals(this.opts.labels.length, 2);
+      assert.equals(this.opts.labels[0], 'foo');
+      assert.equals(this.opts.labels[1], 'bar');
+      done();
+    });
+    cli.exec();
+  }
+});
+
+buster.testCase('cli - ssh', {
+  'should contain ssh command and delegate to breaker ssh when exec is called': function (done) {
+    this.stub(bag, 'cli', {
+      command: function (base, actions) {
+        actions.commands.ssh.action('df -kh');
+      },
+      exit: bag.cli.exit
+    });
+    this.stub(Breaker.prototype, 'ssh', function (command, cb) {
+      assert.equals(command, 'df -kh');
+      assert.equals(typeof cb, 'function');
+      done();
+    });
+    cli.exec();
+  },
+  'should pass labels array via constructor when argument labels are provided': function (done) {
+    this.stub(bag, 'cli', {
+      command: function (base, actions) {
+        actions.commands.ssh.action('df -kh', { labels: 'foo,bar' });
+      },
+      exit: bag.cli.exit
+    });
+    this.stub(Breaker.prototype, 'ssh', function (command, cb) {
+      assert.equals(this.opts.labels.length, 2);
+      assert.equals(this.opts.labels[0], 'foo');
+      assert.equals(this.opts.labels[1], 'bar');
+      done();
+    });
+    cli.exec();
+  }
+});