+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Rename config file from hosts.json to .breaker.json
 * Add config file generation support supporting various formats
 * Remove tasks library support
+* Add label filtering to format and ssh targets
 
 ### 0.0.3
 * Rename project to Sshm (last one!)